Improve tolerance definition.
In sklearn, they set to zeros the values that are very close to zero.
Then they try to perform the power method. If all columns of the
Y matrix are too small, then they consider the algorihtm to have
converged.
This is a very similar approach, but taking into account the
magnitude of X and Y to calculate the epsilons