Tentative of NaN interpolation in the 1D case
It uses the left and right non-NaN values to compute the interpolation
manually. Unfortunately this approach is way slower than calling
np.interp with the appropriate points for each row. However, I still
want to commit this to Git history, as the approach has merit and could
be useful in the future.