Allow to reduce the grid dimension in the DP algorithm for registration
In order for elastic registration to be tractable with a high number of
points, we can now reduce the grid used, so that only the energy of
direct lines from near grid points are computed. This should not alter
much the result (lines with far away points can be approximated with
more line segments), but can drastically improve performance.

We now use trapezoid quadrature to compute the integrals that appear in
the energy computation. The reason for that change is that it is easier
to compute the right quadrature weights when integrals with different
domains are computed in a vectorized way.