Fix karcher mean only iterating twice
A bug was found in the Karcher mean due to the use of the `out=`
parameter: the original mean and the mean of the next iteration were
aliasing the same memory, so their difference was always 0.

This caused the Karcher mean function to potentially exit early, not
doing all the iterations.