Fix: Make line integrals in DP algorithm finer
There was a problem with the accuracy of the line integrals in the DP
algorithm: the times after the warping where the curve change slope are
not guaranteed to correspond with the original ones. Thus, if the
warping compresses the time, the integral could not take into account
many details of the warped function. This was specially noticeable with
non-smooth curves, such as in the Berkely Growth dataset.

As a temporary solution, we increase the number of points used in the
discretization of the curves, by a factor of `dim_grim` (which roughly
corresponfs to the maximum slope). This improves accuracy, but makes
the algorithm slower again.